Close #1
Implemented GET request to fetch all images in teasers list, not just
the first one. These images then had to be compared to the images in
deck, but there were frequently dimension mismatches. To remedy the
dimension mismatches, I implemented a cross-correlation template
matching algorithm on fast Fourier transforms of images in deck and
images in the teasers list. Then, the resulting image of
cross-correlation coefficients is then queried for its maximum
value. If the smaller template is a match for a subsection of the
larger image, the maximum cross-correlation coefficient will be 1. For more information, refer to
`plot_template.ipynb` and the scikit-image documentation for the
function:
https://scikit-image.org/docs/dev/api/skimage.feature.html#skimage.feature.match_template .